requirements.txt
pytesseract
Pillow
pdf2image